+++ doc/CHANGELOG.markdown
@@ -1,6 +1,8 @@
 
-Filez-2.0.0
-*
+Filez-2.0.0 (WIP)
+* Fixed some IE7 rendering bug
+* Ldap user factory now allow to retrieve accounts from baseDn subtrees
+* Fixed PHP-5.3 incompatibilities
 
 Filez-2.0.0-BETA2
 * Added 'password' field to fz_file table. You need to run 'db.migration-filez-2.0.0-1.sql'